--- core/testing/edu/asu/c3simulator/testing/WidgetTest.java
@@ -1,17 +1,0 @@
-package edu.asu.c3simulator.testing;
-
-
-public interface WidgetTest
-{
-	void testResizingZero();
-	
-	void testResizingNegative();
-	
-	void testResizingBoundaryUpper();
-	
-	void testResizingBoundaryLower();
-	
-	void testResizingPrecisionUpper();
-	
-	void testResizingPrecisionLower();
-}
